+++ tests/test_relation_index.py
@@ -29,6 +29,7 @@ from inmanta_module_factory.inmanta.implement import Implement
 from inmanta_module_factory.inmanta.implementation import Implementation
 from inmanta_module_factory.inmanta.index import Index
 from inmanta_module_factory.inmanta.module import Module
+from inmanta_module_factory.inmanta.types import InmantaStringType
 
 LOGGER = logging.getLogger(__name__)
 
@@ -53,7 +54,7 @@ def test_basic(project: Project) -> None:
         fields=[
             Attribute(
                 name="test",
-                inmanta_type="string",
+                inmanta_type=InmantaStringType,
                 description="This is a test attribute",
             ),
         ],
@@ -67,7 +68,7 @@ def test_basic(project: Project) -> None:
         fields=[
             Attribute(
                 name="test",
-                inmanta_type="string",
+                inmanta_type=InmantaStringType,
                 description="This is a test attribute",
             ),
         ],