+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-# v 1.2.0 (?)
+# v 2.0.0 (?)
 Changes in this release:
+ - Added support for typedef statement. (#109)
+ - Refactored the type hierarchy.
 
 # v 1.1.0 (2022-03-18)
 Changes in this release: